feat: Add websocket for alarms
Also, the server is now the one who decides if the alarm should activate for [5]. This is better because
a) The server should be in charge of deciding that.
b) It reduces the number of HTTP calls. In the previous commit, we had "get number of people" + post "alarm". Now it's just post "rpir motion".