feat: Synchronize MyPiEvent access
Since I'm about to introduce setting the D4S7 event in a thread for displaying current time, it occurred to me we need to synchronize MyPiEvent's event setting.
It also occurred to me that we don't need this class at all, and in fact shouldn't have it? Each event getting fired should *immediately* perform the requested event and fire the associated _on_event_func's. We're pointlessly bottlenecking all events (which may be fired in parallel) through this class. Locking would still be necessary, but for each actuator individually, probably in the components module.